backend/register/requirements.txt
Flask
pymongo
werkzeug
flask_cors
pytest
requests
Flask-Mail
python-dotenv
PyJWT